requirements.txt
streamlit==1.30.0
openai==1.5.0
pyautogen==0.2.7
tqdm==4.66.1